Добавлена проверка ссылку на ID, если ID из базы данных не совпадает, то записывается ссылка в базу данных. Дубликаты ссылок в базе данных удаляются.

+++ handlers/handlers.py
@@ -1,5 +1,5 @@
 import re
-
+import sqlite3
 from aiogram import F
 from aiogram import types
 from aiogram.exceptions import TelegramBadRequest
@@ -64,10 +64,47 @@ async def left_chat_member(message: types.Message):
     await message.delete()  # Удаляем системное сообщение о покидании группы
 
 
+async def write_to_database(link):
+    try:
+        # Подключение к базе данных (создаст файл базы данных, если он не существует)
+        conn = sqlite3.connect('setting/database.db')
+        cursor = conn.cursor()
+
+        # Создание таблицы, если она не существует
+        cursor.execute('''
+            CREATE TABLE IF NOT EXISTS links (
+                id INTEGER PRIMARY KEY AUTOINCREMENT,
+                link TEXT NOT NULL UNIQUE,
+                added_on TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+            )
+        ''')
+
+        # Проверка на наличие дубликата
+        cursor.execute('SELECT COUNT(1) FROM links WHERE link = ?', (link,))
+        if cursor.fetchone()[0] > 0:
+            logger.info(f"Ссылка '{link}' уже существует в базе данных. Запись не требуется.")
+            return
+
+        # Вставка новой записи в таблицу
+        cursor.execute('INSERT INTO links (link) VALUES (?)', (link,))
+
+        # Сохранение изменений
+        conn.commit()
+
+        logger.info(f"Ссылка '{link}' успешно добавлена в базу данных.")
+
+    except sqlite3.Error as error:
+        logger.error(f"Ошибка при работе с базой данных: {error}")
+
+    finally:
+        # Закрытие соединения с базой данных
+        conn.close()
+
+
 @dp.message(F.text)
 async def any_message(message: types.Message):
     """
-    Проверка сообщения на наличие ссылок если ссылка в сообщении есть, то удаляем сообщение и предупреждаем пользователя.
+    Проверка сообщения на наличие ссылок: если ссылка в сообщении есть, то удаляем сообщение и предупреждаем пользователя.
     """
     logger.info(f'Проверяем сообщение {message.text} от {message.from_user.username} {message.from_user.id}')
     logger.info(f'Текст сообщения: {message.text}')
@@ -79,30 +116,50 @@ async def any_message(message: types.Message):
                     link = message.text[
                            entity.offset:entity.offset + entity.length] if entity.type != "text_link" else entity.url
                     logger.info(f"Ссылка ({entity.type}) в сообщении 🔗: {link}")
+
+                    # Проверка на наличие ссылки в базе данных
+                    conn = sqlite3.connect('setting/database.db')
+                    cursor = conn.cursor()
+                    cursor.execute('SELECT COUNT(1) FROM links WHERE link = ?', (link,))
+                    link_exists = cursor.fetchone()[0] > 0
+                    conn.close()
+
+                    if link_exists:
+                        logger.info(f"Ссылка '{link}' найдена в базе данных. Проверка по ID не требуется.")
+                        continue  # Переход к следующей ссылке, если она найдена в базе данных
+
+                    logger.info(f"Ссылка '{link}' не найдена в базе данных. Продолжаем проверку по ID.")
                     client, username_id = await connect_session_to_telegram_account(link)
+
                     if message.from_user.id not in allowed_user_ids:
                         logger.info(f'ID группы {link}: {username_id}')
                         users = await read_database()
+                        user_found = False
                         for user in users:
                             logger.info(f'ID из базы данных: {user[0]}')
                             if username_id == user[0]:
+                                user_found = True
                                 user_id = message.from_user.id
                                 permissions = ChatPermissions(can_send_messages=False)
                                 try:
                                     await bot.restrict_chat_member(chat_id=message.chat.id, user_id=user_id,
-                                                               permissions=permissions)
-                                    logger.info(
-                                        f'Сообщение от:({message.from_user.username} {message.from_user.id}). Текст сообщения {message.text}')
+                                                                   permissions=permissions)
+                                    logger.info(f'Сообщение от:({message.from_user.username} '
+                                                f'{message.from_user.id}). Текст сообщения {message.text}')
 
                                     await message.delete()  # Удаляем сообщение содержащее ссылку
                                     await client.disconnect()
                                 except TelegramBadRequest:
                                     await client.disconnect()
-
-                    logger.info(f'Сообщение от админа:({message.from_user.username} {message.from_user.id}). Текст сообщения {message.text}')
+                        if not user_found:
+                            logger.info(
+                                f'ID группы {link}: {username_id} не найдено в базе данных. Записываем ссылку в базу данных.')
+                            await write_to_database(link)  # Запись ссылки в базу данных
+                    logger.info(f'Сообщение от админа:({message.from_user.username} '
+                                f'{message.from_user.id}). Текст сообщения {message.text}')
                     await client.disconnect()
     except Exception as error:
-        logger.info(f'Ошибка: {error}') # Обработка ошибки
+        logger.info(f'Ошибка: {error}')  # Обработка ошибки
 
 
 @dp.edited_message(F.text)
@@ -131,8 +188,8 @@ async def edit_message(message: types.Message):
                             permissions = ChatPermissions(can_send_messages=False)
                             await bot.restrict_chat_member(chat_id=message.chat.id, user_id=user_id,
                                                            permissions=permissions)
-                            logger.info(
-                                f'Сообщение от:({message.from_user.username} {message.from_user.id}). Текст сообщения {message.text}')
+                            logger.info(f'Сообщение от:({message.from_user.username} '
+                                        f'{message.from_user.id}). Текст сообщения {message.text}')
                             await message.delete()  # Удаляем сообщение содержащее ссылку
                             await client.disconnect()
     except Exception as error:
